Converted use of double to long for calculations.

+++ src/payroll/hole09/TaxCalculator.java
@@ -1,5 +1,5 @@
 package payroll.hole09;
 
 public interface TaxCalculator {
-	double taxFor(final double grossSalary);
+	long taxFor(final long grossSalary);
 }
